Use WORD type for GetNumSections() and GetSectionHeader().
The number of sections is a WORD, so the header index can not exceed
that either. This avoids casting from (signed) int to (unsigned) UINT.

R=pkasting@chromium.org

Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3679938
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1009114}

+++ base/win/pe_image.h
@@ -116,7 +116,7 @@ class PEImage {
 
   // Returns the header for a given section.
   // returns NULL if there is no such section.
-  PIMAGE_SECTION_HEADER GetSectionHeader(UINT section) const;
+  PIMAGE_SECTION_HEADER GetSectionHeader(WORD section) const;
 
   // Returns the size of a given directory entry or 0 if |directory| is out of
   // bounds.
